uint32_t

+++ coldaemon.h
@@ -150,7 +150,7 @@ void thread_add(struct thread_list **lista, int index);
 pthread_t * thread_get(struct thread_list *lista, int index);
 void thread_del(struct thread_list **lista, int index);
 uint32_t hash( char * str);
-char authentication (char * acl_file, char * user, unsigned long  pass_buscado);
+char authentication (char * acl_file, char * user, uint32_t pass_buscado);
 
 /*
 	Administrador de Hilos
